add locking to file level
